+++ src/examples/index.js
@@ -1,4 +1,4 @@
-import * as infiniteScroll from './infiniteScroll';
+import * as InfiniteScroll from './InfiniteScroll';
 import * as chat from './chat';
 
-export default { chat, infiniteScroll };
+export default { chat, InfiniteScroll };
